Compute key block length when exporting TR-31 key block
Instead of relying on the length field in the TR-31 context object, the
initial length field in the TR-31 header will be populated as zero and
updated later. It is important that the key block length is updated in
the TR-31 header before the authenticator is generated, otherwise it
will be incorrect.

This change also sets the payload length and authenticator length when
the format version is validated, instead of in the encrypt+sign helper
functions.

Unit tests have also been updated to provide an incorrect length field
in the TR-31 context object.