Add decryption verification to TR-31 export unit test
Although it's not ideal to use tr31_import() to test tr31_export(), at
least tr31_import() itself was easier to verify using the TR-31:2018 test
vectors as well as test vectors from a Thales payShield 10k, so using it
to test tr31_export() isn't entirely flawed.

The primary reason for adding this is to detect regressions in
authenticator generation.