Simplify use of header pointer in TR-31 context object
The purpose of this pointer is convey the header data to the decrypt and
encrypt helper functions. It was simplified by having it point to the
key block buffer instead of allocating and copying to a separate buffer.
This pointer can also be const as it is not used to modify the header
data.