Update optional block TS for ANSI X9.143
ASC TR-31:2018 was fairly vague about the format requirements for
optional block TS while ANSI X9.143 clarifies it.

The intention of this implementation is that the tr31 library expects
the correct ISO 8601 format during both import and export. It is up to
the caller to convert the date/time to and from other formats. In
future, tr31-tool may provide such conversions.