Implement optional block TC for ANSI X9.143
This implementation uses the same helper function for ISO 8601
validation as optional block TS.

The intention of this implementation is that the tr31 library expects
the correct ISO 8601 format during both import and export. It is up to
the caller to convert the date/time to and from other formats. In
future, tr31-tool may provide such conversions.