Create new tr31_strings API
The intention is to make it easier to implement more complex string
rendering (for date/time conversion, certificate parsing, etc)
separately from the primary TR-31 API such that it can easily be omitted
for embedded builds.

For now, only the rendering of human readable descriptions of optional
block data has been moved to tr31_strings. This has also been refactored
to become tr31_opt_block_data_get_desc() for more consistent naming and
to write the output string to a buffer instead of returning a pointer to
a string literal. This allows more complex strings to be rendered.