Implement sscanf() fallback for strptime()
This implementation uses sscanf() when strptime() isn't available. Note
that some struct tm values have to be adjusted after parsing. The
strptime() function is currently not widely available, but may be
available in future if it is standardised as part of ISO C2x. Until
then, an alternative is useful.