Update library tests for ANSI X9.143
Note that not all examples in ANSI X9.143:2021 are correct. Those tests
are either omitted, as in tr31_decrypt_test, or corrected, as in
tr31_export_test. In addition, tr31_export_test has been refactored to
loop over the test data instead of duplicating the code for every test
case.