Refactor and fix key version handling
The initial implementation incorrectly assumed format N for the key
version while the key version should hold format AN instead. This change
corrects that mistake and simplifies the key version in the TR-31
context object. Although this change is not backward compatible, the
intention is for the next release to update the SONAME version anyway.