Only allocate optional block data after validation
No memory leak was possible because tr31_import() calls tr31_release()
upon error regardless. However, in future the decoding of optional block
data will be separated from the validation in which case the allocation
will occur after validation. It is also not necessary to use calloc()
for the cases when data of the exact length will be copied into the
buffer right after allocation, in which case malloc() is sufficient.