Implement tr31_opt_block_find() to lookup optional blocks
The intention is that users of the API either iterate over all optional
blocks in the TR-31 context object or lookup the ones that interest
them. This API function satisfies the latter.

Another benefit of this function is also that unit tests can move the
optional block validation out of their massive if-statements.