Refactor tr31_import() not to require null-termination
It is convenient for the caller to be able to pass a key block that was
received in some kind of message format (eg ISO 8583) directly to
tr31_import() without having to copy it out to add null-termination.
However, this requires an additional parameter to tr31_import() for the
caller to indicate the exact key block length.

Note that this API change is not backward compatible due to the
additional parameter for tr31_import() as well as the renumbering of the
error enum. The intention is for the next release to update the SONAME
version anyway.

In addition, tr31_import() will now also validate the key block buffer
as printable ASCII as early as possible. This ensures that there aren't
any illegal characters or NULL values in the key block buffer.