Persist CMake variables used during multiple builds in cache
When a variable created during configuration is used during multiple
builds, and isn't written to a generated header, it should be a cache
variable. If it isn't a cache variable, it will be unavailable during
subsequent builds that perform partial configuration, which in turn has
unintended consequences for those builds.